更新

+++ 普中51板/SG90舵机实验360/App/sg90/sg90.h
@@ -6,7 +6,7 @@
 //数据线
 sbit PWM = P2^1;
 
-extern signed int time;
+extern u8 time;
 extern u8 time_cnt;
 
 # endif
